+++ packages/css-color-parser/src/functions/rgb-normalize-channel-values.ts
@@ -72,7 +72,10 @@ export function normalize_modern_sRGB_ChannelValues(token: CSSToken, index: numb
 			colorData.syntaxFlags.add(SyntaxFlag.HasPercentageValues);
 		}
 
-		const value = normalize(token[4].value, 100, 0, 1);
+		let value = normalize(token[4].value, 100, -Infinity, Infinity);
+		if (index === 3) {
+			value = normalize(token[4].value, 100, 0, 1);
+		}
 
 		return [
 			TokenType.Number,
@@ -91,7 +94,7 @@ export function normalize_modern_sRGB_ChannelValues(token: CSSToken, index: numb
 			colorData.syntaxFlags.add(SyntaxFlag.HasNumberValues);
 		}
 
-		let value = normalize(token[4].value, 255, 0, 1);
+		let value = normalize(token[4].value, 255, -Infinity, Infinity);
 		if (index === 3) {
 			value = normalize(token[4].value, 1, 0, 1);
 		}